requirements.txt
google-api-python-client
bcolz
matplotlib
numpy
opencv-python
Pillow
torch